+++ aomc/enum_test.go
@@ -1,0 +1,25 @@
+package aomc_test
+
+import (
+	"testing"
+
+	"github.com/philippgille/apiomat-go/aomc"
+)
+
+func TestAuthImplStatus(t *testing.T) {
+	expected := "UNKNOWN"
+	given := aomc.Unknown
+	actual := given.String()
+	if expected != actual {
+		t.Errorf("expected %v, actual %v", expected, actual)
+	}
+}
+
+func TestUserRole(t *testing.T) {
+	expected := "User"
+	given := aomc.User
+	actual := given.String()
+	if expected != actual {
+		t.Errorf("expected %v, actual %v", expected, actual)
+	}
+}